+++ packages/index.ts
@@ -1,2 +1,3 @@
 export { Badge } from "./badge";
 export { BadgeButton } from "./badge";
+export { Column, Table, Cell, TextCell, HeaderCell } from "./table";