+++ packages/table/components/HeaderCell.tsx
@@ -1,0 +1,7 @@
+import { default as Cell } from "./Cell";
+import styled from "react-emotion";
+
+export default styled(Cell)`
+  text-transform: uppercase;
+  font-weight: bold;
+`;