+++ packages/table/index.ts
@@ -1,0 +1,5 @@
+export { default as Table } from "./components/Table";
+export { default as Column } from "./components/Column";
+export { default as Cell } from "./components/Cell";
+export { default as HeaderCell } from "./components/HeaderCell";
+export { default as TextCell } from "./components/TextCell";